change create order logic
